requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.40.0
anyio==4.6.2.post1
arrow==1.3.0
attrs==24.2.0
bidict==0.23.1
binaryornot==0.4.4
blessed==1.20.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cookiecutter==2.6.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
editor==1.6.6
email_validator==2.2.0
eval_type_backport==0.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.28.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
inquirer==3.4.0
Jinja2==3.1.4
jinxed==1.3.0
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.8
langchain-anthropic==0.3.0
langchain-community==0.3.8
langchain-core==0.3.21
langchain-google-genai==2.0.6
langchain-mongodb==0.2.0
langchain-openai==0.2.10
langchain-redis==0.1.1
langchain-text-splitters==0.3.2
langsmith==0.1.146
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
motor==3.6.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.55.1
orjson==3.10.12
packaging==24.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.18.0
pymongo==4.9.2
pyparsing==3.2.0
PyPDF2==3.0.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.17
python-slugify==8.0.4
python-socketio==5.11.4
python-ulid==2.7.0
PyYAML==6.0.2
readchar==4.2.1
redis==5.2.0
redisvl==0.3.6
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
runs==1.2.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
starlette==0.41.3
tabulate==0.9.0
tenacity==9.0.0
text-unidecode==1.3
tiktoken==0.8.0
tldextract==5.1.3
tqdm==4.67.1
typer==0.13.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
wsproto==1.2.0
xmod==1.8.1
yarl==1.18.0
